requirements.txt
# NYC Real Estate AI - Python Dependencies

# Core AI/ML
google-generativeai>=0.8.0  # Gemini Flash for property data extraction

# Database
supabase>=2.27.0  # Supabase PostgreSQL client
psycopg2-binary>=2.9.0  # PostgreSQL adapter (for direct SQL if needed)

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment variable management
python-dotenv>=1.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# NYC Open Data API access (ACRIS, PLUTO)
sodapy>=2.2.0

# Excel file parsing (DOF Rolling Sales)
openpyxl>=3.1.0

# Web UI
streamlit>=1.52.0

# Legacy OpenAI support (optional, if needed)
# openai>=1.0.0

# Legacy Airtable support (optional, for migration)
# pyairtable>=2.1.0
